refactor for packaging
