+++ adept/_lpse2d/core/laser.py
