+++ adept/_lpse2d/core/vector_field.py
@@ -1,10 +1,10 @@
-from typing import Dict, List
+from typing import Dict
 
-from jax import numpy as jnp
+from jax import numpy as jnp, Array
 import numpy as np
 
 from adept import get_envelope
-from adept.lpse2d.core import epw, laser
+from adept._lpse2d.core import epw, laser
 
 
 class SplitStep:
@@ -12,8 +12,6 @@ class SplitStep:
     This class contains the function that updates the state
 
     All the pushers are chosen and initialized here and a single time-step is defined here.
-
-    Note that EPW can be either the charge density (div E) or the potential
 
     :param cfg:
     :return:
@@ -24,7 +22,6 @@ class SplitStep:
         self.cfg = cfg
         self.dt = cfg["grid"]["dt"]
         self.wp0 = cfg["units"]["derived"]["wp0"]
-        # self.epw = epw.FDChargeDensity(cfg)
         self.epw = epw.SpectralPotential(cfg)
         self.light = laser.Light(cfg)
         self.complex_state_vars = ["E0", "epw"]
@@ -35,7 +32,7 @@ class SplitStep:
 
         self.nu_coll = cfg["units"]["derived"]["nu_coll"]
 
-    def unpack_y(self, y):
+    def _unpack_y_(self, y: Dict[str, Array]) -> Dict[str, Array]:
         new_y = {}
         for k in y.keys():
             if k in self.complex_state_vars:
@@ -44,16 +41,23 @@ class SplitStep:
                 new_y[k] = y[k].view(jnp.float64)
         return new_y
 
-    def pack_y(self, y, new_y):
+    def _pack_y_(self, y: Dict[str, Array], new_y: Dict[str, Array]) -> tuple[Dict[str, Array], Dict[str, Array]]:
         for k in y.keys():
             y[k] = y[k].view(jnp.float64)
             new_y[k] = new_y[k].view(jnp.float64)
 
         return y, new_y
 
-    def light_split_step(self, t, y, args):
-        t_coeff = get_envelope(0.03, 0.03, 0.1, 100.0, t)
-        y["E0"] = self.boundary_envelope[..., None] * t_coeff * self.light.laser_update(t, y, args["E0"])
+    def light_split_step(self, t, y, driver_args):
+        if "E0" in driver_args:
+            t_coeff = get_envelope(
+                driver_args["E0"]["tr"],
+                driver_args["E0"]["tr"],
+                driver_args["E0"]["tc"] - driver_args["E0"]["tw"] / 2,
+                driver_args["E0"]["tc"] + driver_args["E0"]["tw"] / 2,
+                t,
+            )
+            y["E0"] = self.boundary_envelope[..., None] * t_coeff * self.light.laser_update(t, y, driver_args["E0"])
 
         # if self.cfg["terms"]["light"]["update"]:
         # y["E0"] = y["E0"] + self.dt * jnp.real(k1_E0)
@@ -65,7 +69,7 @@ class SplitStep:
 
         return y
 
-    def landau_damping(self, epw, vte_sq):
+    def landau_damping(self, epw: Array, vte_sq: float):
         gammaLandauEpw = (
             np.sqrt(np.pi / 8)
             * self.wp0**4
@@ -76,18 +80,20 @@ class SplitStep:
 
         return jnp.fft.ifft2(jnp.fft.fft2(epw) * jnp.exp(-(gammaLandauEpw + self.nu_coll) * self.dt))
 
-    def __call__(self, t: float, y: Dict, args: Dict) -> Dict:
+    def __call__(self, t, y, args):
         # unpack y into complex128
-        new_y = self.unpack_y(y)
+        new_y = self._unpack_y_(y)
 
         # split step
         new_y = self.light_split_step(t, new_y, args["drivers"])
+
+        if "E2" in args["drivers"]:
+            new_y["epw"] += self.dt * self.epw.driver(args["drivers"]["E2"], t)
         new_y["epw"] = self.epw(t, new_y, args)
 
         # landau and collisional damping
-        new_y["epw"] = self.landau_damping(epw=new_y["epw"], vte_sq=y["vte_sq"])
-
-        new_y["epw"] = jnp.fft.ifft2(self.zero_mask * jnp.fft.fft2(new_y["epw"]))
+        if self.cfg["terms"]["epw"]["damping"]["landau"]:
+            new_y["epw"] = self.landau_damping(epw=new_y["epw"], vte_sq=y["vte_sq"])
 
         # boundary damping
         ex, ey = self.epw.calc_fields_from_phi(new_y["epw"])
@@ -95,9 +101,8 @@ class SplitStep:
         ey = ey * self.boundary_envelope
         new_y["epw"] = self.epw.calc_phi_from_fields(ex, ey)
         new_y["epw"] = jnp.fft.ifft2(self.zero_mask * self.low_pass_filter * jnp.fft.fft2(new_y["epw"]))
-        # new_y["epw"] = new_y["epw"] * self.boundary_envelope
 
         # pack y into float64
-        y, new_y = self.pack_y(y, new_y)
+        y, new_y = self._pack_y_(y, new_y)
 
         return new_y
