+++ tests/test_vlasov1d/test_absorbing_wave.py
@@ -8,7 +8,7 @@ config.update("jax_enable_x64", True)
 from jax import jit
 from diffrax import ODETerm, diffeqsolve
 
-from adept.vlasov1d.solvers.pushers.field import Driver, WaveSolver
+from adept._vlasov1d.solvers.pushers.field import Driver, WaveSolver
 from adept import Stepper
